Chatbot/requirements.txt
stremlit
google-genereativeal
python-dotenv
